+++ ckanext/superset/auth/superset_database.py
@@ -1,0 +1,5 @@
+
+
+def superset_database_list(context, data_dict):
+    """ Sysadmin only """
+    return {'success': False}