+++ ckanext/superset/plugin.py
@@ -3,6 +3,8 @@ from ckan import plugins
 from ckan.plugins import toolkit
 from ckanext.superset.actions import superset_dataset as superset_dataset_actions
 from ckanext.superset.auth import superset_dataset as superset_dataset_auth
+from ckanext.superset.actions import superset_database as superset_database_actions
+from ckanext.superset.auth import superset_database as superset_database_auth
 
 
 log = logging.getLogger(__name__)
@@ -24,12 +26,14 @@ class SupersetPlugin(plugins.SingletonPlugin):
 
     def get_actions(self):
         return {
-            "superset_dataset_list": superset_dataset_actions.superset_dataset_list
+            "superset_dataset_list": superset_dataset_actions.superset_dataset_list,
+            "superset_database_list": superset_database_actions.superset_database_list
         }
 
     # IAuthFunctions
 
     def get_auth_functions(self):
         return {
-            "superset_dataset_list": superset_dataset_auth.superset_dataset_list
+            "superset_dataset_list": superset_dataset_auth.superset_dataset_list,
+            "superset_database_list": superset_database_auth.superset_database_list
         }