+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 Features:
  - New API endpoint at `/api/action/superset_dataset_list` to get data from Superset URL at `/api/v1/datasets/`
    [#1](https://github.com/unckan/ckanext-superset/pull/1)
+
+- New API endpoint at `/api/action/superset_database_list` to get data from Superset URL at `/api/v1/database/`
+  [#2](https://github.com/unckan/ckanext-superset/pull/2)